+++ src/components/form/text-field/types.ts
@@ -9,4 +9,7 @@ export interface TextFieldProps {
     placeholder?: string
     value?: string
     icon?: IconNameType
+    actionIcon?: IconNameType
+    onActionTriggered?: () => void
+    isPassword?: boolean
 }