+++ src/components/index.ts
@@ -1,7 +1,8 @@
 import Button from './button'
 import Icon from './icon'
+import PasswordField from './form/password-field'
 import Text from './text'
 import Title from './title'
 import TextField from './form/text-field'
 
-export { Button, Icon, Text, TextField, Title }
+export { Button, Icon, PasswordField, Text, TextField, Title }